#430 - Introduce JPA managed type access to EntityViewProxy and move to SPI. Base equals on JPA managed type to fix #430
