+++ entity-view/impl/src/main/java/com/blazebit/persistence/view/impl/accessor/ViewIdAttributeAccessor.java
@@ -18,7 +18,7 @@ package com.blazebit.persistence.view.impl.accessor;
 
 import com.blazebit.persistence.view.impl.EntityViewManagerImpl;
 import com.blazebit.persistence.view.impl.update.UpdateContext;
-import com.blazebit.persistence.view.impl.proxy.EntityViewProxy;
+import com.blazebit.persistence.view.spi.type.EntityViewProxy;
 import com.blazebit.persistence.view.metamodel.ViewType;
 
 /**
