--- entity-view/impl/src/main/java/com/blazebit/persistence/view/impl/proxy/EntityViewProxy.java
@@ -1,31 +1,0 @@
-/*
- * Copyright 2014 - 2017 Blazebit.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.blazebit.persistence.view.impl.proxy;
-
-@SuppressWarnings("checkstyle:methodname")
-public interface EntityViewProxy {
-
-    public boolean $$_isNew();
-
-    public void $$_setIsNew(boolean isNew);
-    
-    public Class<?> $$_getEntityViewClass();
-
-    public Object $$_getId();
-    
-    public Object $$_getVersion();
-}
