+++ entity-view/testsuite/src/test/java/com/blazebit/persistence/view/testsuite/proxy/ProxyFactoryTest.java
@@ -242,8 +242,8 @@ public class ProxyFactoryTest extends AbstractEntityViewTest {
 
         // 5 Fields
         assertEquals(5, proxyClass.getDeclaredFields().length);
-        // 5 Getters, 2 Setter, 1 Bridge-Getter, 1 Bridge-Setter, 1 Equals, 1 HashCode, 5 EntityViewProxy methods
-        assertEquals(16, proxyClass.getDeclaredMethods().length);
+        // 5 Getters, 2 Setter, 1 Bridge-Getter, 1 Bridge-Setter, 1 Equals, 1 HashCode, 6 EntityViewProxy methods
+        assertEquals(17, proxyClass.getDeclaredMethods().length);
         assertAttribute(proxyClass, "contacts", Modifier.PRIVATE, Map.class, Integer.class, Person.class);
         assertAttribute(proxyClass, "myContactPerson", Modifier.PRIVATE | Modifier.FINAL, Person.class);
         assertAttribute(proxyClass, "firstContactPerson", Modifier.PRIVATE | Modifier.FINAL, Person.class);
@@ -262,8 +262,8 @@ public class ProxyFactoryTest extends AbstractEntityViewTest {
 
         // 5 Fields
         assertEquals(5, proxyClass.getDeclaredFields().length);
-        // 5 Getters, 2 Setter, 1 Bridge-Getter, 1 Bridge-Setter, 1 Equals, 1 HashCode, 5 EntityViewProxy methods
-        assertEquals(16, proxyClass.getDeclaredMethods().length);
+        // 5 Getters, 2 Setter, 1 Bridge-Getter, 1 Bridge-Setter, 1 Equals, 1 HashCode, 6 EntityViewProxy methods
+        assertEquals(17, proxyClass.getDeclaredMethods().length);
         assertAttribute(proxyClass, "contacts", Modifier.PRIVATE, Map.class, Integer.class, Person.class);
         assertAttribute(proxyClass, "myContactPerson", Modifier.PRIVATE | Modifier.FINAL, Person.class);
         assertAttribute(proxyClass, "firstContactPerson", Modifier.PRIVATE | Modifier.FINAL, Person.class);
